% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add a package to the packages.R file}
\usage{
add_package(pkgs, homedir = ".", bioC = FALSE)
}
\description{
Add a package to the packages.R file
}
\examples{
add_packages(c('tidyverse','broom','stringr'))
}
