% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useTemplate.R
\name{useTemplate}
\alias{useTemplate}
\title{useTemplate}
\usage{
useTemplate(homedir = ".")
}
\arguments{
\item{homedir}{What directory are you creating the template in}
}
\description{
This function actually creates the folders and files that go into the template
}
